Exception handling - delete

+++ src/main/java/com/lucasweb/demo/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.lucasweb.demo.entities.User;
 import com.lucasweb.demo.repositories.UserRepository;
+import com.lucasweb.demo.services.exception.DatabaseExeption;
 import com.lucasweb.demo.services.exception.ResourceNotFoundException;
 
 @Service
@@ -29,10 +32,17 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);
+	public void delete(Long id){
+	    try {
+	        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+	        repository.deleteById(id);
+	    }catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    } catch(DataIntegrityViolationException e) {
+	    	throw new DatabaseExeption(e.getMessage());
+	    }
 	}
-	
+	 
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
